[visualizations] lazy load actions (#207147)
Clean up visualizations page load size by
* lazy loading actions
* avoid exporting from index files to avoid exporting unused code
* move `urlFor` and `getFullPath` into `url_utils` to avoid including
`utils/saved_visualize_utils` in page load bundle

---------

(cherry picked from commit d7f801ab3e9c22cc77c0fc3da9492008a5a5252a)
