+++ packages/kbn-optimizer/limits.yml
@@ -196,5 +196,5 @@ pageLoadAssetSize:
   visTypeVega: 153573
   visTypeVislib: 242838
   visTypeXy: 46868
-  visualizations: 90000
+  visualizations: 41000
   watcher: 43598
