+++ src/platform/plugins/shared/visualizations/public/actions/edit_in_lens_action.tsx
@@ -42,8 +42,7 @@ import {
   getUsageCollection,
 } from '../services';
 import { DASHBOARD_VISUALIZATION_PANEL_TRIGGER } from '../triggers';
-
-export const ACTION_EDIT_IN_LENS = 'ACTION_EDIT_IN_LENS';
+import { ACTION_EDIT_IN_LENS } from './constants';
 
 const displayName = i18n.translate('visualizations.actions.editInLens.displayName', {
   defaultMessage: 'Convert to Lens',
